requirements.txt
Django==1.11.1
django-crispy-forms==1.6.1
django-cors-headers==2.0.2
django-material
django-debug-toolbar
django-extensions>=1.5.9
django-avatar
django-redis==4.8.0
pymysql==0.7.11
urllib3==1.21.1
colorlog
Werkzeug
celery==4.0.2
flower
sphinx
sphinx_rtd_theme
swagger2rst
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
coreapi
PyJWT>=1.4.0,<2.0.0
pytz
requests
Pillow
Faker

gunicorn
